Adding ssh-keys to the release configurations.
Clean up unused lines of command.

The ssh-agent configuration was tested successfully on a full cycle of Indigo Trust 64 with the keys removed from the slaves. There are other jobs that will not work without the keys. In particular the ssh-agent does not appear to penetrate inside the jenkins_tools chroots.